Update Songs
Update API Hit - 10

+++ routes/songs.js
@@ -30,14 +30,25 @@ router.post('/addplaylist', async(req,res) => {
     }
     else
     {
-        let tambahPlaylist = `insert into h_playlist values ('','${resultUser[0].id_user}','${user.username}','${nama_playlsit}')`;
-        let resPlaylist = await dbase.executeQuery(tambahPlaylist);
-
-        temp = {
-            "nama_playlist" : nama_playlsit,
-            "ditambahkan_oleh" : resultUser[0].nama_user
+        if(resultUser[0].api_hit - 10 >= 0)
+        {
+            let tambahPlaylist = `insert into h_playlist values ('','${resultUser[0].id_user}','${user.username}','${nama_playlsit}')`;
+            let resPlaylist = await dbase.executeQuery(tambahPlaylist);
+            
+            let apiupdate = resultUser[0].api_hit-10;
+            let updateApi = `update users set api_hit=${apiupdate} where username='${user.username}'`;
+            let resUpdate = await dbase.executeQuery(updateApi);
+    
+            temp = {
+                "nama_playlist" : nama_playlsit,
+                "ditambahkan_oleh" : resultUser[0].nama_user
+            }
+            return res.status(201).send(temp);
         }
-        return res.status(201).send(temp);
+        else
+        {
+            return res.status(400).send("API Hit tidak mencukupi. Silahkan melakukan pengisian saldo untuk membeli API Hit");
+        }
     }
 });
 
@@ -57,22 +68,46 @@ router.get('/title', async(req,res) => {
     if((new Date().getTime()/1000)-user.iat>10*60){//10mnt
         return res.status(400).send("Token expired");
     }
-    var arrSong = [];
-    let querySearch = 'https://api.genius.com/search?access_token=' + token + '&q=' + req.query.name;
-    try {
-        let resultGet = await axios.get(querySearch);
-        console.log(resultGet.data.response.hits.length);
-        for (let i = 0; i < resultGet.data.response.hits.length; i++) {
-            temp = {
-                "songs_id" : resultGet.data.response.hits[i].result.id,
-                "songs_name" : resultGet.data.response.hits[i].result.full_title
-            }
-            arrSong.push(temp);
-        }
-        return res.status(200).send(arrSong);
-    } catch (error) {
-        return res.status(404).send("Lagu tidak ditemukan");
+
+    let cekUser = `select * from users where username='${user.username}'`;
+    let resultUser = await dbase.executeQuery(cekUser);
+    if(resultUser.length < 1)
+    {
+        return res.status(404).send("User tidak ditemukan");
     }
+    else
+    {
+        if(resultUser[0].api_hit - 10 >= 0)
+        {
+            var arrSong = [];
+            let querySearch = 'https://api.genius.com/search?access_token=' + token + '&q=' + req.query.name;
+            try {
+                let resultGet = await axios.get(querySearch);
+                console.log(resultGet.data.response.hits.length);
+                for (let i = 0; i < resultGet.data.response.hits.length; i++) {
+                    temp = {
+                        "songs_id" : resultGet.data.response.hits[i].result.id,
+                        "songs_name" : resultGet.data.response.hits[i].result.full_title
+                    }
+                    arrSong.push(temp);
+                }
+
+                let apiupdate = resultUser[0].api_hit-10;
+                let updateApi = `update users set api_hit=${apiupdate} where username='${user.username}'`;
+                let resUpdate = await dbase.executeQuery(updateApi);
+                
+                return res.status(200).send(arrSong);
+            } catch (error) {
+                return res.status(404).send("Lagu tidak ditemukan");
+            }
+        }
+        else
+        {
+            return res.status(400).send("API Hit tidak mencukupi. Silahkan melakukan pengisian saldo untuk membeli API Hit");
+        }
+    }
+
+    
 });
 
 router.post('/addsongtoplaylist', async(req,res) => {
@@ -92,18 +127,43 @@ router.post('/addsongtoplaylist', async(req,res) => {
     let id_playlist = req.body.id_playlist;
     let songs_id = req.body.songs_id;
 
-    let querySearch = 'https://api.genius.com/songs/'+ songs_id +'?access_token=' + token;
-    try {
-        let resultGet = await axios.get(querySearch);
-        let insertSong = `insert into d_playlist values ('${id_playlist}', '${songs_id}','${resultGet.data.response.song.full_title}')`;
-        let resSong = await dbase.executeQuery(insertSong);
-        temp = {
-            
-        }
-        return res.status(201).send("Lagu berhasil ditambahkan ke playlist");
-    } catch (error) {
-        return res.status(404).send("Lagu tidak ditemukan");
+    let cekUser = `select * from users where username='${user.username}'`;
+    let resultUser = await dbase.executeQuery(cekUser);
+    if(resultUser.length < 1)
+    {
+        return res.status(404).send("User tidak ditemukan");
     }
+    else
+    {
+        if(resultUser[0].api_hit - 10 >= 0)
+        {
+            let querySearch = 'https://api.genius.com/songs/'+ songs_id +'?access_token=' + token;
+            try {
+                let resultGet = await axios.get(querySearch);
+                let insertSong = `insert into d_playlist values ('${id_playlist}', '${songs_id}','${resultGet.data.response.song.full_title}')`;
+                let resSong = await dbase.executeQuery(insertSong);
+                temp = {
+                    
+                }
+                let apiupdate = resultUser[0].api_hit-10;
+                let updateApi = `update users set api_hit=${apiupdate} where username='${user.username}'`;
+                let resUpdate = await dbase.executeQuery(updateApi);
+
+                return res.status(201).send("Lagu berhasil ditambahkan ke playlist");
+            } catch (error) {
+                return res.status(404).send("Lagu tidak ditemukan");
+            }
+        }
+        else
+        {
+            return res.status(400).send("API Hit tidak mencukupi. Silahkan melakukan pengisian saldo untuk membeli API Hit");
+        }
+    }
+
+    
+    
+
+    
 });
 
 //Delete playlist
@@ -122,14 +182,36 @@ router.delete('/deleteplaylist', async(req,res) => {
         return res.status(400).send("Token expired");
     }
 
-    let id_playlist = req.body.id_playlist;
-    let deletedplaylist = `delete from d_playlist where id_playlist = '${id_playlist}'`;
-    let resultDel = await dbase.executeQuery(deleteplaylist);
+    let cekUser = `select * from users where username='${user.username}'`;
+    let resultUser = await dbase.executeQuery(cekUser);
+    if(resultUser.length < 1)
+    {
+        return res.status(404).send("User tidak ditemukan");
+    }
+    else
+    {
+        if(resultUser[0].api_hit - 10 >= 0)
+        {
+            let id_playlist = req.body.id_playlist;
+            let deletedplaylist = `delete from d_playlist where id_playlist = '${id_playlist}'`;
+            let resultDel = await dbase.executeQuery(deleteplaylist);
+        
+            let deletehplaylist = `delete from h_playlist where id_playlist = '${id_playlist}'`;
+            let resultDel1 = await dbase.executeQuery(deletehplaylist);
 
-    let deletehplaylist = `delete from h_playlist where id_playlist = '${id_playlist}'`;
-    let resultDel1 = await dbase.executeQuery(deletehplaylist);
+            let apiupdate = resultUser[0].api_hit-10;
+            let updateApi = `update users set api_hit=${apiupdate} where username='${user.username}'`;
+            let resUpdate = await dbase.executeQuery(updateApi);
+        
+            return res.status(200).send("Playlist berhasil dihapus");
+        }
+        else
+        {
+            return res.status(400).send("API Hit tidak mencukupi. Silahkan melakukan pengisian saldo untuk membeli API Hit");
+        }
+    }
 
-    return res.status(200).send("Playlist berhasil dihapus");
+    
 
 });
 
@@ -149,13 +231,35 @@ router.delete('/deletesong', async(req,res) => {
         return res.status(400).send("Token expired");
     }
 
-    let id_playlist = req.body.id_playlist;
-    let songs_id = req.body.songs_id;
+    let cekUser = `select * from users where username='${user.username}'`;
+    let resultUser = await dbase.executeQuery(cekUser);
+    if(resultUser.length < 1)
+    {
+        return res.status(404).send("User tidak ditemukan");
+    }
+    else
+    {
+        if(resultUser[0].api_hit - 10 >= 0)
+        {
+            let id_playlist = req.body.id_playlist;
+            let songs_id = req.body.songs_id;
+        
+            let deleteSong = `delete from d_playlist where id_playlist = '${id_playlist}' and id_lagu='${songs_id}'`;
+            let resultDel = await dbase.executeQuery(deleteSong);
 
-    let deleteSong = `delete from d_playlist where id_playlist = '${id_playlist}' and id_lagu='${songs_id}'`;
-    let resultDel = await dbase.executeQuery(deleteSong);
+            let apiupdate = resultUser[0].api_hit-10;
+            let updateApi = `update users set api_hit=${apiupdate} where username='${user.username}'`;
+            let resUpdate = await dbase.executeQuery(updateApi);
+        
+            return res.status(200).send("Lagu berhasil dihapus dari playlist");
+        }
+        else
+        {
+            return res.status(400).send("API Hit tidak mencukupi. Silahkan melakukan pengisian saldo untuk membeli API Hit");
+        }
+    }
 
-    return res.status(200).send("Lagu berhasil dihapus dari playlist");
+    
 });
 
 //Search lagu di playlist
@@ -176,25 +280,49 @@ router.get('/listSongPlaylist', async(req,res) => {
         return res.status(400).send("Token expired");
     }
     
-    let id_playlist = req.query.id_playlist;
-
-    let querySongPlaylist = `select * from d_playlist where id_playlist='${id_playlist}'`;
-    let resSP = await dbase.executeQuery(querySongPlaylist);
-    if(resSP.length < 1)
+    let cekUser = `select * from users where username='${user.username}'`;
+    let resultUser = await dbase.executeQuery(cekUser);
+    if(resultUser.length < 1)
     {
-        return res.status(404).send("Tidak ada lagu dalam playlist");
+        return res.status(404).send("User tidak ditemukan");
     }
     else
     {
-        arrSong = [];
-        for (let i = 0; i < resSP.length; i++) {
-            temp = {
-                "song_title": resSP[i].title_song
-            };         
-            arrSong.push(temp);   
+        if(resultUser[0].api_hit - 10 >= 0)
+        {
+            let id_playlist = req.query.id_playlist;
+
+            let querySongPlaylist = `select * from d_playlist where id_playlist='${id_playlist}'`;
+            let resSP = await dbase.executeQuery(querySongPlaylist);
+            
+            if(resSP.length < 1)
+            {
+                return res.status(404).send("Tidak ada lagu dalam playlist");
+            }
+            else
+            {
+                arrSong = [];
+                for (let i = 0; i < resSP.length; i++) {
+                    temp = {
+                        "song_title": resSP[i].title_song
+                    };         
+                    arrSong.push(temp);   
+                }
+                let apiupdate = resultUser[0].api_hit-10;
+                let updateApi = `update users set api_hit=${apiupdate} where username='${user.username}'`;
+                let resUpdate = await dbase.executeQuery(updateApi);
+
+                return res.status(200).send(arrSong);
+            }
+            
         }
-        return res.status(200).send(arrSong);
+        else
+        {
+            return res.status(400).send("API Hit tidak mencukupi. Silahkan melakukan pengisian saldo untuk membeli API Hit");
+        }
     }
+
+    
 });
 
 //List Playlist yang dimiliki user sendiri
@@ -213,25 +341,49 @@ router.get('/listPlaylist', async(req,res) => {
         return res.status(400).send("Token expired");
     }
 
-    let playlist = `select * from h_playlist where username_user='${user.username}'`;
-    let resPlay = await dbase.executeQuery(playlist);
-
-    if(resPlay.length < 1)
+    let cekUser = `select * from users where username='${user.username}'`;
+    let resultUser = await dbase.executeQuery(cekUser);
+    if(resultUser.length < 1)
     {
-        return res.status(404).send("User tidak memiliki playlist");
+        return res.status(404).send("User tidak ditemukan");
     }
     else
     {
-        arrPlay = [];
-        for (let i = 0; i < resPlay.length; i++) {
-            temp = {
-                "id_playlist" : resPlay[i].id_playlist,
-                "nama_playlist": resPlay[i].nama_playlist
-            };
-            arrPlay.push(temp);
+        if(resultUser[0].api_hit - 10 >= 0)
+        {
+            let playlist = `select * from h_playlist where username_user='${user.username}'`;
+            let resPlay = await dbase.executeQuery(playlist);
+        
+            if(resPlay.length < 1)
+            {
+                let apiupdate = resultUser[0].api_hit-10;
+                let updateApi = `update users set api_hit=${apiupdate} where username='${user.username}'`;
+                let resUpdate = await dbase.executeQuery(updateApi);
+                return res.status(404).send("User tidak memiliki playlist");
+            }
+            else
+            {
+                arrPlay = [];
+                for (let i = 0; i < resPlay.length; i++) {
+                    temp = {
+                        "id_playlist" : resPlay[i].id_playlist,
+                        "nama_playlist": resPlay[i].nama_playlist
+                    };
+                    arrPlay.push(temp);
+                }
+                let apiupdate = resultUser[0].api_hit-10;
+                let updateApi = `update users set api_hit=${apiupdate} where username='${user.username}'`;
+                let resUpdate = await dbase.executeQuery(updateApi);
+                return res.status(200).send(arrPlay);
+            }
         }
-        return res.status(200).send(arrPlay);
+        else
+        {
+            return res.status(400).send("API Hit tidak mencukupi. Silahkan melakukan pengisian saldo untuk membeli API Hit");
+        }
     }
+
+    
 });
 
 //List Playlist yang dimiliki user lain
